1.4

+++ db/db.go
@@ -3,7 +3,6 @@ package db
 import (
 	"database/sql"
 	"errors"
-	"fmt"
 	"time"
 
 	"golang.org/x/crypto/bcrypt"
@@ -15,6 +14,7 @@ const connStr = "user=postgres password=7458 dbname=test_db sslmode=disable"
 
 var ErrInvalidPassword = errors.New("invalid password")
 var ErrUserNotFound = errors.New("user not found")
+var ErrNoRows = errors.New("no rows")
 
 type response struct {
 	Tasks []data_task `json:"Tasks"`
@@ -111,45 +111,54 @@ func GetAllTasks(id any) (response, error) {
 	data := response{Tasks: tasks}
 	return data, nil
 }
-
-func GetTask(id any) ([]data_task, error) {
+func GetTask(id string, user_id float64) ([]data_task, error) {
 	db, err := sql.Open("postgres", connStr) //Подклчается к БД и проверяем ее
 	if err != nil {
 		panic(err)
 	}
 	defer db.Close()
-	row := db.QueryRow("SELECT tasks.id, status, tasks.name, description, created_at,deadline_at FROM tasks,users WHERE tasks.id = $1", id) // Запрос к БД
+	row := db.QueryRow("SELECT tasks.id, status, tasks.name, description, created_at,deadline_at FROM tasks,users WHERE tasks.id = $1 AND user_id= $2", id, user_id) // Запрос к БД
 	var p data_task
 	err = row.Scan(&p.ID, &p.Status, &p.Name, &p.Description, &p.CreatedAt, &p.Deadline_at)
 	if err != nil {
 		return nil, err
 	}
-	tasks := make([]data_task, 0, 5)
+
+	tasks := make([]data_task, 0, 1) // Создаем слайс с capacitty 1
 	tasks = append(tasks, p)
 	return tasks, nil
 }
 
-func DeleteTask(id any) (string, error) {
+// Надо сделать так что бы удалять данные администратор
+func DeleteTask(id string, user_id float64) (string, error) {
 	db, err := sql.Open("postgres", connStr) //Подклчается к БД и проверяем ее
+	user_id_int := int(user_id)
 	if err != nil {
 		panic(err)
 	}
 	defer db.Close()
-	result, err := db.Exec("DELETE FROM tasks WHERE tasks.id = $1", id)
+	result, err := db.Exec("DELETE FROM tasks WHERE tasks.id = $1 AND tasks.user_id= $2", id, user_id_int)
 	if err != nil {
 		return "", err
 	}
-	fmt.Println(result)
+	rowsAffected, err := result.RowsAffected()
+	if err != nil {
+		return "", err
+	}
+	if rowsAffected == 0 {
+		return "", ErrNoRows
+	}
 	return "", nil
 }
 
-func ChangeTusk(id any, status string) ([]data_task, error) {
+// Надо сделать так что бы менять данные администратор
+func ChangeTusk(id string, status string, user_id float64) ([]data_task, error) {
 	db, err := sql.Open("postgres", connStr) //Подклчается к БД и проверяем ее
 	if err != nil {
 		panic(err)
 	}
 	defer db.Close()
-	_, err = db.Exec("UPDATE tasks SET status = $1 where id = $2", status, id) // Запрос к БД
+	_, err = db.Exec("UPDATE tasks SET status = $1 WHERE id = $2 AND user_id= $3", status, id, user_id) // Запрос к БД
 	if err != nil {
 		return nil, err
 	}
